+++ views/motion/create_select_statutes.php
@@ -1,19 +1,24 @@
 <?php
 
 use app\components\UrlHelper;
-use app\models\db\ConsultationMotionType;
+use app\models\db\{ConsultationAgendaItem, ConsultationMotionType};
 use yii\helpers\Html;
 
 /**
  * @var Yii\web\View $this
  * @var ConsultationMotionType $motionType
+ * @var ConsultationAgendaItem|null $agendaItem
  */
 
 /** @var \app\controllers\Base $controller */
 $controller = $this->context;
 $layout = $controller->layoutParams;
 
-$this->title = $motionType->titleSingular;
+if ($agendaItem) {
+    $this->title = $agendaItem->title . ': ' . $motionType->titleSingular;
+} else {
+    $this->title = $motionType->titleSingular;
+}
 $layout->robotsNoindex = true;
 $layout->addBreadcrumb($motionType->titleSingular);
 
@@ -34,7 +39,11 @@ $layout->addBreadcrumb($motionType->titleSingular);
     $statutes = $motionType->getAmendableOnlyMotions(true, true);
     foreach ($statutes as $statute) {
         echo '<div class="statute statute' . $statute->id . '">';
-        $url = UrlHelper::createUrl(['/amendment/create', 'motionSlug' => $statute->getMotionSlug()]);
+        $urlParams = ['/amendment/create', 'motionSlug' => $statute->getMotionSlug()];
+        if ($agendaItem) {
+            $urlParams['agendaItemId'] = $agendaItem->id;
+        }
+        $url = UrlHelper::createUrl($urlParams);
         $title = '<span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span> ' . Html::encode($statute->getTitleWithPrefix());
         echo Html::a($title, $url);
         echo '</div>';