+++ History.md
@@ -9,6 +9,7 @@
 - A new amendment numbering scheme was introduced more appropriate for english environments: "M1 A1" (Motion number + Amendment number)
 - It is now possible to copy a motion including all its amendments to a different consultation or agenda item without marking the original one as moved.
 - To prevent brute force login attempts, a entering a CAPTCHA is now required after three failed login or account recovery attempts. Standalone hosted versions of Antragsgrün can optionally require it for every single login attempt, by adding the loginCaptcha flag to config.json.
+- Statute amendments can now be created for an agenda item if the according motion type was set.
 - Bugfix: If an amendment was assigned to an agenda item, the agenda item could not be deleted before the amendment was un-assigned first.
 - Bugfix: an empty "Supporting" section was shown in motions and amendments, if only the "liking" function was enabled.
 - Support for Internet Explorer was dropped.